Add feature Max Grid Size to colorplot
Solve memory issues by adding a feature to colorplot: max grid size.
z values will be interpolated on an x-y mesh.
If max gird size absent, the mesh on each axis will be as large as the number of uinque values on the axis, which can become prohibitvely large.
Cap the number of points Z data will be interpolated to by entering max values in the coloplot dialog.